Add files via upload
Switched order of two lines in a query with many joins.  Resolved and issue.  Query result is one row with columns from several tables.  Some of the columns have the same name because certain tables in the database have columns names that are the same (ie. country_name, city_name...).  This was an easier fix than making major changes to the query or making schema changes.